Adds profiling statements and improves slow code. Removes some unused code.
Splits the environment and model data in two different managed library expressions
so that the license is checked only once when the GUROBIEnvironment data is created
and left to be reused through the entire session. The GUROBI model data is kept in
GUROBIData and is created and destroyed with each example.